Moved DevModeInitializerEndpointTest to vaadin-dev-server
